+++ libs/designsystem/divider/src/divider.component.scss
