+++ libs/designsystem/divider/src/public_api.ts
@@ -1,0 +1,1 @@
+export * from './divider.component';
