+++ libs/designsystem/src/lib/components/index.ts
@@ -15,8 +15,6 @@ export { CheckboxComponent } from './checkbox/checkbox.component';
 
 export * from './data-table';
 
-export { DividerComponent } from './divider/divider.component';
-
 export * from './form-field';
 export * from './dropdown';
 export * from './empty-state';
