+++ libs/designsystem/src/lib/index.ts
@@ -20,6 +20,7 @@ export * from '@kirbydesign/designsystem/slide-button';
 export * from '@kirbydesign/designsystem/range';
 export * from '@kirbydesign/designsystem/popover';
 export * from '@kirbydesign/designsystem/loading-overlay';
+export * from '@kirbydesign/designsystem/divider';
 
 export * from './scss/scss-helper';
