+++ libs/designsystem/tsconfig.spec.json
@@ -25,7 +25,8 @@
       "@kirbydesign/designsystem/slide-button": ["slide-button/index.ts"],
       "@kirbydesign/designsystem/range": ["range/index.ts"],
       "@kirbydesign/designsystem/popover": ["popover/index.ts"],
-      "@kirbydesign/designsystem/loading-overlay": ["loading-overlay/index.ts"]
+      "@kirbydesign/designsystem/loading-overlay": ["loading-overlay/index.ts"],
+      "@kirbydesign/designsystem/divider": ["divider/index.ts"]
     }
   },
   "files": ["test.ts"],
